fix(Button): updates button styles to new configuration
